+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [N.N.N] - XXXX-XX-XX
 
+- Polish simulation
+
 ## [0.7.0] - 2026-09-19
 
 ### Interface